Make ready for graalvm native docker images

+++ json-client/src/main/java/com/capgemini/perf/json/client/service/CustomerService.java
@@ -7,5 +7,5 @@ import java.util.Optional;
 public interface CustomerService {
     Iterable<CustomerDTO> all();
 
-    Optional<CustomerDTO> find(Long id);
+    Optional<CustomerDTO> find(int id);
 }
