+++ reference/src/main/java/com/capgemini/perf/reference/service/CustomerService.java
@@ -8,5 +8,5 @@ import java.util.Optional;
 public interface CustomerService {
     Iterable<Customer> all();
 
-    Optional<Customer> find(Long id);
+    Optional<Customer> find(int id);
 }
